Fixed extraneous failover message
Can get cycles.  To test, use Baran Distributed Grid, then in order
break links li, lr, ls, and lF.  (Set blueprint.tooltips=true to see
the link and node IDs).  You’ll see a cycle involving nodes o, n, and t.

The fix coming in a future version is to not assume using a failover
link automatically guarantees success.  Instead, we need to check if
the proposed parent is on the broken branch.