Appears to be a correct version
Found the error that led to hung messages.  The link ack was going to
the wrong side of the link.

Renamed flow control code routine that handles multiple services per
node to entangle2 because it only works if each service has a unique
name.  Added a different version under the name entangle that only
works for one service per node, which is all we have right now.

Fixed a problem in the failover success/failure code that used a single
failover requester.  It needed to be a stack.