Fixed cycle in failover
The trie information was being updated in the wrong order.  In
particular, a node was getting trie information about a non-parent link
of its parent before getting that info from the parent link.  The
result is that it thought the parent wasn’t on a broken branch when it
was.  The fix added a new parameter to the rediscover message,
isParentBranch.  If true, a node will process a rediscover message even
if it has already seen one on this port for this tree.