Failover Working
Fixed several bugs.  One involved how cycles in failover are handled.
Another was how state is reset in order to handle the next failure.  A
third was the labeling of the first element of the path following a
failover that reaches the root.  Also added a count of how many times
each link appears in a tree.