New demo behavior
In this version you select a tree to view by clicking on the node.  You
break a link or a node by double clicking.

This version still has the infinite loop bug.  Don’t break cell “t”
during a demo, but others may cause an infinite loop.  Kill the tab if
that happens.