Improvements to UI and algorithm bug fixes
There are still bugs in the cell breaking code, but the behavior is
much better than the previous version.  A small number of cells on a
small number of trees don’t get updated properly when breaking a cell,
and there is still some infinite sending of rediscover messages, but
more rarely than the previous version.

When you double click on a link, the selected tree stays selected.  The
tree also stays selected with you click Discover Wiring.

I changed the linkIDs to be the nodeIDs of the flanking cells, which
helps in debugging.  I also updated the storyboard to reflect the new
naming and to give precise instructions for a sequence that doesn’t hit
the existing bugs.