Fixed Parent-Child inconsistency bug
There were a few cases where the parent didn’t list the child as a
child.  The problem came from asking my current parent to be my new
parent.  While that doesn’t seem to make sense, it does in the case of
overlapping failures when the current parent has already found a new
path to the root, but the current child doesn’t know it.  The bug was
telling the old parent to drop the child after telling the new parent
to add the child.  Obviously, those requests are in the wrong order if
the old parent is the same as the new parent.

This version also fixes the simultaneous failover request bug except
for the case where the only path back to the root for one requester
goes through the other.