Node Failure Working
This version handles both individual and overlapping link failures.  It
properly handles the case of symmetric failover requests as long as one
of the pair has an alternate path to the root.  It does not handle the
case where neither of them does because the only solution would involve
an infinite loop.  That’s not a problem in a truly distributed system,
but it would lock up this single threaded simulation.